+++ src/helpers.php
@@ -18,15 +18,3 @@ if (! function_exists('getModelForGuard')) {
             })->get($guard);
     }
 }
-
-if (! function_exists('isNotLumen')) {
-    /**
-     * check if application is lumen.
-     *
-     * @return bool
-     */
-    function isNotLumen(): bool
-    {
-        return ! preg_match('/lumen/i', app()->version());
-    }
-}