+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `laravel-permission` will be documented in this file
 
+## 3.7.0 - 2020-02-15
+- Added methods to check any/all when querying direct permissions #1245
+- Removed older Lumen dependencies #1371
+
 ## 3.6.0 - 2020-01-17
 - Added Laravel 7.0 support
 - Allow splat operator for passing roles to `hasAnyRole()`