+++ docs/changelog.md
@@ -1,6 +1,6 @@
 ---
 title: Changelog
-weight: 6
+weight: 10
 ---
 
 All notable changes to laravel-permission are documented [on GitHub](https://github.com/spatie/laravel-permission/blob/master/CHANGELOG.md)